backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
paho-mqtt==1.6.1
